coverage.rs: re-style as scroll list view list_entry.rs: created by abstracting functions out of relay_entry.rs
